read content from theme folder
