+++ VirtoCommerce.Storefront.Model/WorkContext.cs
@@ -155,11 +155,6 @@ namespace VirtoCommerce.Storefront.Model
         public IMutablePagedList<ContentItem> Pages { get; set; }
 
         /// <summary>
-        /// All static content items (Pages, blog articles etc) for current store and theme
-        /// </summary>
-        public IMutablePagedList<ContentItem> PagesInTheme { get; set; }
-
-        /// <summary>
         /// All static templates items (catalog, cart etc) for current store and theme
         /// </summary>
         public IMutablePagedList<ContentItem> Templates { get; set; }
