+++ VirtoCommerce.Storefront/Middleware/WorkContextBuildMiddleware.cs
@@ -56,9 +56,7 @@ namespace VirtoCommerce.Storefront.Middleware
             await builder.WithMenuLinksAsync(workContext.CurrentStore, workContext.CurrentLanguage);
             await builder.WithPagesAsync(workContext.CurrentStore, workContext.CurrentLanguage);
             await builder.WithBlogsAsync(workContext.CurrentStore, workContext.CurrentLanguage);
-
-            await builder.WithPagesInThemeAsync(workContext.CurrentStore, workContext.CurrentLanguage);
-            await builder.WithTemplatesInThemeAsync(workContext.CurrentStore);
+            await builder.WithTemplatesAsync(workContext.CurrentStore);
 
 
             //EU General Data Protection Regulation (GDPR) support
