+++ VirtoCommerce.Storefront/Startup.cs
@@ -86,8 +86,6 @@ namespace VirtoCommerce.Storefront
             services.AddSingleton<IMemberService, MemberService>();
             services.AddSingleton<ISeoInfoService, SeoInfoServise>();
 
-            services.AddSingleton<IStaticContentInThemeService, StaticContentInThemeService>();
-
             services.AddSingleton<IStaticContentService, StaticContentService>();
             services.AddSingleton<IMenuLinkListService, MenuLinkListServiceImpl>();
             services.AddSingleton<IStaticContentItemFactory, StaticContentItemFactory>();
